services/diffusion-engine/requirements.txt
fastapi==0.110.0
uvicorn==0.27.1
torch==2.2.1
diffusers==0.26.3
transformers==4.38.2
accelerate==0.27.2
safetensors==0.4.2
pillow==10.2.0
python-multipart==0.0.9
boto3==1.34.51
pydantic==2.6.1
huggingface-hub==0.25.2
opencv-python-headless==4.9.0.80
numpy==1.26.4
openai-whisper
edge-tts
moviepy==1.0.3
sacremoses
sentencepiece
requests
